% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{abhac.brief}
\alias{abhac.brief}
\title{AbHAC Based on Vector Inputs
This function performs AbHAC analysis on inputs of mutations, upregulated and downregulated proteins}
\usage{
abhac.brief(de.up = NULL, de.down = NULL, snv = NULL,
  ppi.database = NULL, enrichment.categories = NULL, fac = NULL,
  fisher.fdr = "Permutation.FDR", fisher.fdr.cutoff = 0.05,
  id.conversion.set = NULL, num.permuted.ppi, method.permuted.ppi,
  bins.permuted.ppi, num.cores)
}
\arguments{
\item{de.up}{either ensembl gene IDs, HGNC, ENTREZ or uniprot IDs for a vector of upregulated genes}

\item{de.down}{either ensembl gene IDs, HGNC, ENTREZ or uniprot IDs for a vector of downregulated genes}

\item{snv}{either ensembl gene IDs, HGNC, ENTREZ or uniprot IDs for a vector of mutated genes}

\item{ppi.database}{2 column whole protein interaction network. Either loaded by data(ppi.database)(filtering is recommended based on types of interactions) or by user.}

\item{enrichment.categories}{can be all or any of the c("snv.de.up.de.down","de.up","de.down","de","snv","snv.de.up","snv.de.down","snv.de")}

\item{fac}{is all the proteins that exist in protein interaction network. If not using data(ppi.database), it is necessary to specify.}

\item{fisher.fdr}{Can be either "Permutation.FDR", "Permutation.FWER" or any of the methods parsed into p.adjust. Type ?p.adjust for more details.}

\item{fisher.fdr.cutoff}{Cutoff used for false discovery rate cutoff in fisher's exact test. By default set to 0.2.}

\item{id.conversion.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}

\item{num.permuted.ppi}{If you have selected any of the two permutation based methods, the number of permuted networks to be used for multiple testing correction must be specified.}

\item{method.permuted.ppi}{If you have selected any of the two permutation based methods, the method for binning proteins by their edge degree for creting permuted networks for multiple testing correction must be specified. It should be one of ("AsPaper", "equal", "ByDegree").}

\item{bins.permuted.ppi}{If you have selected any of the two permutation based methods, specified the number of bins for proteins to be grouped into. If you have selected "AsPaper", you would better leave this as 4 (default). For the two other methods, we advise a number between 10-20.}

\item{num.cores}{Note that a parallel for loop using foreach package calculates the p-values for all the different permuted networks. The number of processors to be used for this foreach has to be set by you using the "registerDoMC(cores=4)". However this parameters determined the number of processors to be used for calculation of pvalues by an mclapply feature. So if you are using registerDoMC(cores=4) and you want to limit the analysis to 12 processors, you must specify num.cores = 3.}
}
\value{
This function returns a data.frame with results of integrative network enrichment analysis
}
\description{
AbHAC Based on Vector Inputs
This function performs AbHAC analysis on inputs of mutations, upregulated and downregulated proteins
}
\examples{
data(snv)
data(rna)
snv = sample(rownames(snv),10)
de.up = sample(rownames(rna)[1:1000],500)
de.down = sample(rownames(rna)[1001:2000],500)
data(ppi.database) #2column whole human protein interaction database
data(id.conversion.set)
data(fac) #vector of all proteins in ppi.database
abhac.brief.result = abhac.brief(de.up,de.down,fac=fac,snv=snv,enrichment.categories=c("snv.de","de.up"),ppi.database=ppi.database[,1:2],id.conversion.set=id.conversion.set)
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

