% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cgna.brief}
\alias{cgna.brief}
\title{CGNA Based on Vector Inputs
This function performs CGNA analysis on inputs of mutations, upregulated and downregulated proteins}
\usage{
cgna.brief(de.up = NULL, de.down = NULL, snv = NULL,
  ppi.database = NULL, enrichment.categories = NULL, fac = NULL,
  fisher.fdr = "Permutation", fisher.fdr.cutoff = 0.2,
  id.conversion.set = NULL)
}
\arguments{
\item{de.up}{either ensembl gene IDs, HGNC, ENTREZ or uniprot IDs for a vector of upregulated genes}

\item{de.down}{either ensembl gene IDs, HGNC, ENTREZ or uniprot IDs for a vector of downregulated genes}

\item{snv}{either ensembl gene IDs, HGNC, ENTREZ or uniprot IDs for a vector of mutated genes}

\item{ppi.database}{2 column whole protein interaction network. Either loaded by data(ppi.database)(filtering is recommended based on types of interactions) or by user.}

\item{fac}{is all the proteins that exist in protein interaction network. If not using data(ppi.database), it is necessary to specify.}

\item{enrichment.categories}{can be all or any of the c("snv.de.up.de.down","de.up","de.down","de","snv","snv.de.up","snv.de.down","snv.de")}

\item{id.conversion.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}

\item{fisher.fdr}{Can be either "Permutation" or any of the methods parsed into p.adjust. Type ?p.adjust for more details.}

\item{fisher.fdr.cutoff}{Cutoff used for false discovery rate cutoff in fisher's exact test. By default set to 0.2.}
}
\value{
This function returns a data.frame with results of integrative network enrichment analysis
}
\description{
CGNA Based on Vector Inputs
This function performs CGNA analysis on inputs of mutations, upregulated and downregulated proteins
}
\examples{
data(snv)
data(rna)
snv = sample(rownames(snv),10)
de.up = sample(rownames(rna)[1:1000],500)
de.down = sample(rownames(rna)[1001:2000],500)
data(ppi.database) #2column whole human protein interaction database
data(id.conversion.set)
data(fac) #vector of all proteins in ppi.database
cgna.brief.result = cgna.brief(de.up,de.down,fac=fac,snv=snv,enrichment.categories=c("snv.de","de.up"),ppi.database=ppi.database[,1:2],id.conversion.set=id.conversion.set)
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

