% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{clinical}
\alias{clinical}
\title{TCGA BReast Cancer clinical information table for 40 patients}
\format{First column represent patients in same order as mutation matrix and expression matrix. Subsequent columns provide some clinical information.}
\source{
TCGA
}
\usage{
clinical
}
\description{
This data set is obtained From TCGA Breast Cancer Dataset and is incomplete and only used as an example to run functions
}
\author{
Mehran Karimzadeh
}

