% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{deizer}
\alias{deizer}
\title{Differential Expression calculator function}
\usage{
deizer(rna = NULL, expression.method = NULL, i = NULL, paired = NULL,
  correction.method = NULL, fdr.cutoff = NULL, fac = NULL)
}
\arguments{
\item{rna}{is expression matrix, method of determined expression, and index for samples to be included and pvalue adjustment method paired, a boolean, indicate if for each sample, we should expect adjacent normal tissue or not}

\item{expression.method}{is a string either "Microarray" or "RNAseq"}

\item{i}{is index of columns to calculate differential expression agains columns having N}

\item{paired}{is a boolean showing if RNA samples are paired}

\item{correction.method}{Any of the "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" or "fdr" can be used.}

\item{fdr.cutoff}{0.01 or 0.05 are suggested. All values in renge of 0 and 1 are accepted}

\item{fac}{is all the proteins that exist in protein interaction network. If not using data(ppi.database), it is necessary to specify.}
}
\value{
a list with 3 string vectors: de.up, de.down and de
}
\description{
Differential Expression calculator function
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

