% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{ensembl.to.uniprot}
\alias{ensembl.to.uniprot}
\title{Ensembl Gene ID to Uniprot Accession Conversion}
\usage{
ensembl.to.uniprot(ensembl = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{ensembl}{is a vector of ensembl gene IDs}

\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}
}
\value{
If a Uniprot accession is not found, it returns the ID itself instead of ""
}
\description{
Converts ENSEMBL Human Gene IDs to Uniprot
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

