% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{entrez.to.uniprot}
\alias{entrez.to.uniprot}
\title{Entrez Gene ID to Uniprot accession}
\usage{
entrez.to.uniprot(entrez = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}

\item{Entrez}{is vector of Entrez ID characters}
}
\value{
If a Entrez accession is not found, the return will be the ID itself
}
\description{
Converts Entrez Gene ID to uniprot accessions
}
\examples{
entrez.to.uniprot(c("12", "5421", "65223"))
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

