% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{fac}
\alias{fac}
\title{Proteins Annotated in ppi.database data}
\format{Unipriot Accession}
\source{
TCGA
}
\usage{
fac
}
\description{
This is simply a vector of all proteins that an interaction for them exists in ppi.database dataset
}
\author{
Mehran Karimzadeh
}

