% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{get_edgedegree_freq}
\alias{get_edgedegree_freq}
\title{Making Protein Edge Reoccurence dataframe}
\usage{
get_edgedegree_freq(ppi.database, fac)
}
\arguments{
\item{ppi.database}{2 column undirected protein interaction network.}

\item{fac}{vector of unique proteins inside ppi.database}
}
\value{
A dataframe describing frequency of proteins within each edge degree.
}
\description{
Used inside the fisher exact test motor function for creating subsequent permuted networks.
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

