% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hgnc.to.uniprot}
\alias{hgnc.to.uniprot}
\title{HGNC Symbol to Uniprot Accession Conversion}
\usage{
hgnc.to.uniprot(hgnc = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{hgnc}{is one HGNC symbol accession string}

\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}
}
\value{
If a Uniprot accession is not found, it returns the ID itself instead of ""
}
\description{
Converts human gene symbols to uniprot swissprot acession
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

