% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{id.conversion.set}
\alias{id.conversion.set}
\title{Table for ID Conversions inside fucntions}
\format{Columns are Entrez Gene ID, Uniprot Acession, Human Gene Symbols with Aliases seperated by space, Human Ensembl Gene IDs and Human Protein Refseq IDs}
\source{
UniProt.ss
}
\usage{
id.conversion.set
}
\description{
This data set is obtained with UniProt.ws bioconductor package getting all entrez gene IDs and their associated uniprot IDs, ensembl gene IDs, Human Gene Symbols and Human Protein Refseq IDs
}
\author{
Mehran Karimzadeh
}

