% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{Integrator}
\alias{Integrator}
\title{AbHAC Internal Enrichment Calculator}
\usage{
Integrator(ppi.database = NULL, list.categories = NULL, fac = NULL,
  id.conversion.set = NULL, fisher.fdr = "Permutation.FDR",
  fisher.fdr.cutoff = 0.05, num.permuted.ppi = 10,
  method.permuted.ppi = "AsPaper", bins.permuted.ppi = 4, num.cores = 6)
}
\arguments{
\item{ppi.database}{2 column whole protein interaction network. Either loaded by data(ppi.database)(filtering is recommended based on types of interactions) or bu used.}

\item{list.categories}{(Internal) is list of proteins for each enrichment category with accurate names as of enrichment.categories}

\item{fac}{is all the proteins that exist in protein interaction network. If not using data(ppi.database), it is necessary to specify.}

\item{id.conversion.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}

\item{fisher.fdr}{Can be either "Permutation.FDR", "Permutation.FWER" or any of the methods parsed into p.adjust. Type ?p.adjust for more details.}

\item{fisher.fdr.cutoff}{Cutoff to be used for fisher's exact test false discovery rates.}

\item{num.permuted.ppi}{If you have selected any of the two permutation based methods, the number of permuted networks to be used for multiple testing correction must be specified.}

\item{method.permuted.ppi}{If you have selected any of the two permutation based methods, the method for binning proteins by their edge degree for creting permuted networks for multiple testing correction must be specified. It should be one of ("AsPaper", "equal", "ByDegree").}

\item{bins.permuted.ppi}{If you have selected any of the two permutation based methods, specified the number of bins for proteins to be grouped into. If you have selected "AsPaper", you would better leave this as 4 (default). For the two other methods, we advise a number between 10-20.}

\item{num.cores}{Note that a parallel for loop using foreach package calculates the p-values for all the different permuted networks. The number of processors to be used for this foreach has to be set by you using the "registerDoMC(cores=4)". However this parameters determined the number of processors to be used for calculation of pvalues by an mclapply feature. So if you are using registerDoMC(cores=4) and you want to limit the analysis to 12 processors, you must specify num.cores = 3.}
}
\value{
A dataframe with Benjamini-Hochberg FDR bases on fisher's one tail exact test for all enrichment categories for all proteins that interact with at least one of the input proteins
}
\description{
Internal Function that performs abhac analysis inside other functions
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

