% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{multiple.testing.correction.handler}
\alias{multiple.testing.correction.handler}
\title{Corrects p-values by the provided multiple testing correction method}
\usage{
multiple.testing.correction.handler(list.pvalues, fisher.fdr, fisher.fdr.cutoff)
}
\arguments{
\item{list.pvalues}{2 column dataframes of protein names and associated p-values must be put in a list structure where the 1st list member corresponds to the p-values based on the true protein interaction network and other list members if they exist must be based on permuted interaction networks}

\item{fisher.fdr}{Multiple testing correction method from: "Permutation.FDR", "Permutation.FWER", or p.adjust methods.}

\item{fisher.fdr.cutoff}{the cutoff for FWER of FDR (numeric)}
}
\description{
Corrects p-values by the provided multiple testing correction method
}
\author{
Mehran Karimzadeh mehran dot karimzadehreghbati at mail dot mcgill dot ca
}

