% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{permute_pdr}
\alias{permute_pdr}
\title{Create a permuted network}
\usage{
permute_pdr(ppi.database, df_pr_freq, method = "AsPaper", k = 4,
  verbose = FALSE)
}
\arguments{
\item{ppi.database}{2 column undirected protein interaction network}

\item{df_pr_freq}{dataframe of first column as unique protein names of ppi.database, and 2nd column as number of times they have occured in the network. This is calculated by get_edgedegree_freq}

\item{method}{Specify "equal" for equal number of proteins in each bin sorted by edge degree, "ByDegree" to create bins solely based on edge degree and "AsPaper" to have one bin for each edge degree unless the number of proteins in the edge degree is lower than 3. For such proteins create k equal size bins.}

\item{k}{number of bins to divide the proteins and permute inside those bins}
}
\value{
ppi_permuted 2 column undirected protein interaction network
}
\description{
Used inside the fisher exact test motor function, it will provide permuted networks for estimating FDR/FWER.
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

