% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ppi.database}
\alias{ppi.database}
\title{Whole Human Protein Interaction Network}
\format{First two columns as proteins interacting with each other, and subsequenc columns representing information}
\source{
PSICQUIC
}
\usage{
ppi.database
}
\description{
This data set is obtained with bioconductor PSICQUIC package accessing following databases: DIP,InnateDB,IntAct,MatrixDB,MINT,I2D-IMEx,InnateDB-IMEx,MolCon,BindingDB
}
\author{
Mehran Karimzadeh
}

