% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Random.ppis}
\alias{Random.ppis}
\title{10 randomized Protein interaction networks}
\format{A list representing 10 randomized networks in format of 2 column dataframes representing undirected interaction information}
\source{
PSICQUIC Randomized
}
\usage{
Random.ppis
}
\description{
This data set is obtained with bioconductor PSICQUIC package accessing following databases: DIP,InnateDB,IntAct,MatrixDB,MINT,I2D-IMEx,InnateDB-IMEx,MolCon,BindingDB and then randomized 10 times and represented as a list
}
\author{
Mehran Karimzadeh
}

