% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{refseqp.to.uniprot}
\alias{refseqp.to.uniprot}
\title{Refseq Protein to Uniprot Conversion Function}
\usage{
refseqp.to.uniprot(refseqp = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{refseqp}{is one Human Protein Refseq ID accession string}

\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}
}
\value{
If a Uniprot accession is not found, it returns the ID itself instead of ""
}
\description{
Converts Human  Protein Refseq IDs to Uniprot
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

