% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{remove_duplicate_interactions}
\alias{remove_duplicate_interactions}
\title{Keeps only one undirected edge between two proteins}
\usage{
remove_duplicate_interactions(ppi.database)
}
\arguments{
\item{ppi.database}{2 column undirected protein interaction network}
}
\value{
A protein interaction 2 column dataframe of unique undirected interactions
}
\description{
Keeps only one undirected edge between two proteins
}
\author{
Mehran Karimzadeh mehran dot karimzadehreghbati at mail dot mcgill dot ca
}

