% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{rna.id.conversion}
\alias{rna.id.conversion}
\title{Expression Matrix ID Conversion}
\usage{
rna.id.conversion(rna = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{rna}{for this function is a matrix of expression with rownames as genes and colnames as samples}

\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}
}
\value{
This function finds uniprot IDs, removes rows without uniprot ID or duplicate uniprot ID while averaging values
}
\description{
Converts non-uniprot rownames of expression matrix to uniprot. merges rows that output rowname is same.(by averaging)
IDs should be in one of these formats: entrez gene IDs, HGNC or ENSEMBL gene IDs
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

