% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\docType{data}
\name{rna}
\alias{rna}
\title{TCGA BReast Cancer Expression Matrix for 40 Tumors and 22 Nontumor breast tissues}
\format{rownames represent gene names as uniprot accession and each column is one patient. Values are obtained from normalized Agilient Microarray. Only top 2000 most variant genes are shown. Column names with T inicate Tumor and with N indicate Nontumor. First 40 patients are in same order as mutation matrix.}
\source{
TCGA
}
\usage{
rna
}
\description{
This data set is obtained From TCGA Breast Cancer Dataset and is incomplete and only used as an example to run functions
}
\author{
Mehran Karimzadeh
}

