% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{set.abhac}
\alias{set.abhac}
\title{This Function performs integrative analysis on the whole dataset}
\usage{
set.abhac(ppi.database = NULL, rna = NULL, snv = NULL,
  expression.method = NULL, rna.paired = FALSE, correction.method = "BH",
  fdr.cutoff = 0.05, enrichment.categories = NULL,
  id.conversion.set = NULL, fac = NULL, clinical = NULL,
  fisher.fdr = "Permutation", fisher.fdr.cutoff = 0.05,
  num.permuted.ppi = 10, bins.permuted.ppi = 4,
  method.permuted.ppi = "AsPaper", num.cores = 6)
}
\arguments{
\item{ppi.database}{2 column whole protein interaction network. Either loaded by data(ppi.database)(filtering is recommended based on types of interactions) or by user.}

\item{rna}{a numeric matrix with rownames either as gene symbols, ensemble genes IDs, entrez IDs or uniprot IDs and columns representing same patient IDs as mutation matrix but with T and N suffices standing for Tumor and Nontumor samples.}

\item{snv}{a character matrix with row names as either of uniprot IDs, gene symbols, entrez IDs or ensemble gene Ids. When there is no mutation, NA is expected.}

\item{expression.method}{"RNAseq" or "Microarray"}

\item{rna.paired}{boolean indicating if RNA samples are paired}

\item{correction.method}{for differential expression analysis "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" or "fdr"}

\item{fdr.cutoff}{values higher than 0.1 are not advised}

\item{enrichment.categories}{can be all or any of the c("snv.de.up.de.down","de.up","de.down","de","snv","snv.de.up","snv.de.down","snv.de")}

\item{id.conversion.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}

\item{fac}{is all the proteins that exist in protein interaction network. If not using data(ppi.database), it is necessary to specify.}

\item{fisher.fdr}{Can be either "Permutation" or any of the methods parsed into p.adjust. Type ?p.adjust for more details.}

\item{fisher.fdr.cutoff}{Cutoff used for false discovery rate cutoff in fisher's exact test. By default set to 0.2.}

\item{num.permuted.ppi}{If you have selected any of the two permutation based methods, the number of permuted networks to be used for multiple testing correction must be specified.}

\item{bins.permuted.ppi}{If you have selected any of the two permutation based methods, specified the number of bins for proteins to be grouped into. If you have selected "AsPaper", you would better leave this as 4 (default). For the two other methods, we advise a number between 10-20.}

\item{method.permuted.ppi}{If you have selected any of the two permutation based methods, the method for binning proteins by their edge degree for creting permuted networks for multiple testing correction must be specified. It should be one of ("AsPaper", "equal", "ByDegree").}

\item{num.cores}{Note that a parallel for loop using foreach package calculates the p-values for all the different permuted networks. The number of processors to be used for this foreach has to be set by you using the "registerDoMC(cores=4)". However this parameters determined the number of processors to be used for calculation of pvalues by an mclapply feature. So if you are using registerDoMC(cores=4) and you want to limit the analysis to 12 processors, you must specify num.cores = 3.}
}
\value{
This function returns a data.frame with results of integrative network enrichment analysis
}
\description{
This Function performs integrative analysis on the whole dataset
}
\examples{
data(snv)
data(rna)
data(ppi.database) #2column whole human protein interaction database
data(id.conversion.set)
data(fac) #vector of all proteins in ppi.database
set.abhac.result = set.abhac(snv=snv,rna=rna,fac=fac,expression.method="Microarray",rna.paired=FALSE,
   fdr.cutoff=0.05,correction.method="BH",enrichment.categories=c("snv.de","de.up"),
   ppi.database=ppi.database[,1:2],id.conversion.set=id.conversion.set)
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

