% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{snv.id.conversion}
\alias{snv.id.conversion}
\title{Mutation Matrix ID Conversion}
\usage{
snv.id.conversion(snv = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{snv}{for this function is a matrix of mutations with rownames as genes and colnames as samples}

\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}
}
\value{
This function finds uniprot IDs, removes rows without uniprot ID or duplicate uniprot ID
}
\description{
Converts non-uniprot rownames of mutation matrix to uniprot. merges rows that output rowname is same.
IDs should be in one of these formats: entrez gene IDs, HGNC or ENSEMBL gene IDs
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

