% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\name{uniprot.to.entrez}
\alias{uniprot.to.entrez}
\title{Uniprot Accession to Entrez Gene ID conversion}
\usage{
uniprot.to.entrez(uniprot = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{uniprot}{is a vector of uniprot accession IDs}

\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}
}
\value{
If a Entrez accession is not found, it returns the ID itself instead of "".
}
\description{
Converts Uniprot accession to Entrez Gene ID
}
\examples{
uniprot.to.entrez(ids.to.uniprot("VHL"))
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

