% Generated by roxygen2 (4.0.1): do not edit by hand
\name{uniprot.to.hgnc}
\alias{uniprot.to.hgnc}
\title{Uniprot Accession to HGC symbol Conversion}
\usage{
uniprot.to.hgnc(uniprot = NULL, id.con.set = id.conversion.set)
}
\arguments{
\item{uniprot}{is one Uniprot accession string}

\item{id.con.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}
}
\value{
If a HGNC is not found, it returns the ID itself instead of ""
}
\description{
This function is used internally to convert uniprot Accession to HGNC symbol
}
\examples{
uniprot.to.hgnc(ids.to.uniprot("VHL"))
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

