% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\docType{data}
\name{fac}
\alias{fac}
\title{Proteins Annotated in ppi.database data}
\format{Unipriot Accession}
\source{
TCGA
}
\usage{
fac
}
\description{
This is simply a vector of all proteins that an interaction for them exists in ppi.database dataset
}
\author{
Mehran Karimzadeh
}

