% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Integrator}
\alias{Integrator}
\title{CGNA Internal Enrichment Calculator}
\usage{
Integrator(ppi.database = NULL, list.categories = NULL, fac = NULL,
  id.conversion.set = NULL, fisher.fdr = "Permutation",
  fisher.fdr.cutoff = 0.2)
}
\arguments{
\item{ppi.database}{2 column whole protein interaction network. Either loaded by data(ppi.database)(filtering is recommended based on types of interactions) or bu used.}

\item{list.categories}{(Internal) is list of proteins for each enrichment category with accurate names as of enrichment.categories}

\item{fac}{is all the proteins that exist in protein interaction network. If not using data(ppi.database), it is necessary to specify.}

\item{id.conversion.set}{A dataframe for ID conversions provided as global variable id.conversion.set. Columns represent Entrez gene ID, Uniprot Accession, Gene Symbol, Ensembl gene ID and refseq protein ID (all human)}

\item{fisher.fdr}{Can be either "Permutation" or any of the methods parsed into p.adjust. Type ?p.adjust for more details.}

\item{fisher.fdr.cutoff}{Cutoff to be used for fisher's exact test false discovery rates.}
}
\value{
A dataframe with Benjamini-Hochberg FDR bases on fisher's one tail exact test for all enrichment categories for all proteins that interact with at least one of the input proteins
}
\description{
Internal Function that performs cgna analysis inside other functions
}
\author{
Mehran Karimzadeh mehran.karimzadehreghbati at mail dot mcgill dot ca
}

