% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbHAC.R
\docType{data}
\name{snv}
\alias{snv}
\title{TCGA BReast Cancer Mutation Matrix for 40 Patients}
\format{rownames represent gene names as uniprot accession and each column is one patient. Values of NA or "Mutation" indicate status of each patient for each gene}
\source{
TCGA
}
\usage{
snv
}
\description{
This data set is obtained From TCGA Breast Cancer Dataset and is incomplete and only used as an example to run functions
}
\author{
Mehran Karimzadeh
}

